pyproject.toml
[tool.poetry]
name = "nlp-ua-locations-extractions"
version = "0.1.0"
description = ""
authors = ["EvgeniyAzarov <azarov.evg.ua@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tqdm = "^4.66.1"
transformers = "^4.35.2"
pandas = "^2.1.3"
spacy = "^3.7.2"
ipykernel = "^6.26.0"
ipywidgets = "^8.1.1"
torch = "^2.1.1"
lingua-language-detector = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
